+++ src/components/PageNotFound/PageNotFound.js
@@ -1,19 +1,28 @@
 import { Link } from 'react-router-dom';
 import { MAIN } from '../../utils/routesMap';
 import { forPageNotFound as config } from '../../configs/configForComponents';
-import './PageNoutFound.css';
+import './PageNotFound.css';
 
-function PageNoutFound() {
+/**
+ * @module PageNotFound
+ * @description Функциональный компонент<br>
+ * Блок с сообщением, что пользователь пытался перейти на несуществующую страницу.
+ * @returns {JSX}
+ * @since v.1.1.0
+ */
+function PageNotFound() {
   const { TITLE, DESCRIPTION, ERROR_STATUS, LINK_TEXT } = config;
   return (
     <section className="page-not-found">
       <div className="page-not-found__element"></div>
-      <h2 className="page-not-found__title">{ERROR_STATUS}</h2>
-      <h3 className="page-not-found__title">{TITLE}</h3>
+      <span className="page-not-found__status">{ERROR_STATUS}</span>
+      <h2 className="page-not-found__title">{TITLE}</h2>
       <p className="page-not-found__description">{DESCRIPTION}</p>
-      <Link to={MAIN}>{LINK_TEXT}</Link>
+      <Link className="page-not-found__link" to={MAIN}>
+        {LINK_TEXT}
+      </Link>
     </section>
   );
 }
 
-export default PageNoutFound;
+export default PageNotFound;
