+++ src/utils/MainApi.js
@@ -2,6 +2,22 @@ import { BASE_URL } from '../configs';
 import * as PATH_TO from './endpoints';
 import { getTokenFromStorage } from './storage';
 
+/**
+ * @module MainApi
+ * @description API взаимодействия пользователя со статьями.
+ * @since v.1.1.0
+ */
+
+/**
+ * @method register
+ * @description Регистрация пользователя, принимает регистрационные данные пользователя и отправляет
+ *  запрос на регистрацию на сервер.
+ * @param {String} email - емэйл
+ * @param {String} password - пароль
+ * @param {String} name - имя
+ * @returns {JSON}
+ * @since v.1.1.0
+ */
 export const register = (email, password, name) => {
   return fetch(`${BASE_URL}${PATH_TO.REGISTER}`, {
     method: 'POST',
@@ -20,6 +36,15 @@ export const register = (email, password, name) => {
     });
 };
 
+/**
+ * @method login
+ * @description Авторизация пользователя, принимает авторизационные данные пользователя и отправляет
+ *  запрос на авторизацию на сервер.
+ * @param {String} email - емэйл
+ * @param {String} password - пароль
+ * @returns {JSON}
+ * @since v.1.1.0
+ */
 export const login = (email, password) => {
   return fetch(`${BASE_URL}${PATH_TO.LOGIN}`, {
     method: 'POST',
@@ -32,6 +57,12 @@ export const login = (email, password) => {
     });
 };
 
+/**
+ * @method getUserDataFromDataBase
+ * @description Метод получения данных пользователя из базы данных
+ * @returns {JSON}
+ * @since v.1.1.0
+ */
 export const getUserDataFromDataBase = () => {
   return fetch(`${BASE_URL}${PATH_TO.USER}`, {
     method: 'GET',
@@ -46,6 +77,12 @@ export const getUserDataFromDataBase = () => {
     });
 };
 
+/**
+ * @method getSavedNewsFromDataBase
+ * @description Метод получения из БД данных о сохраненных пользователем новостях
+ * @returns {JSON}
+ * @since v.1.1.0
+ */
 export const getSavedNewsFromDataBase = () => {
   return fetch(`${BASE_URL}${PATH_TO.SAVED_NEWS}`, {
     method: 'GET',
@@ -60,6 +97,13 @@ export const getSavedNewsFromDataBase = () => {
     });
 };
 
+/**
+ * @method deleteArticle
+ * @description Метод удаления статей из БД.
+ * @param {String} articleId - id  удаляемой статьи
+ * @returns {JSON}
+ * @since v.1.1.0
+ */
 export const deleteArticle = (articleId) => {
   return fetch(`${BASE_URL}${PATH_TO.SAVED_NEWS}/${articleId}`, {
     method: 'DELETE',
@@ -75,6 +119,13 @@ export const deleteArticle = (articleId) => {
   });
 };
 
+/**
+ * @method addArticleToSavedNews
+ * @description Метод сохранения статьи в БД.
+ * @param {Object} article - объект с данными статьи
+ * @returns {JSON}
+ * @since v. 1.1.0
+ */
 export const addArticleToSavedNews = (article) => {
   return fetch(`${BASE_URL}${PATH_TO.SAVED_NEWS}`, {
     method: 'POST',
