+++ src/utils/date.js
@@ -1,3 +1,11 @@
+/**
+ * @function getFormattedDate
+ * @description  Утилита преобразования даты.<br>
+ * Принимает дату в формате UTC, преобразует в строку типа "2 января, 2020"
+ * @param {String} dateUTC -  строка, дата статьи
+ * @returns {String}
+ * @since v.1.1.0
+ */
 export const getFormattedDate = (dateUTC) => {
   const months = [
     'января',
